Update CentralNicReseller transfer() and register() functions
